refactor: refatora domínio Patients
