+++ apps/apae/src/app/patients/[id]/edit/[step]/page.tsx
@@ -1,17 +1,17 @@
 "use client";
-import { useParams, useRouter } from "next/navigation"; 
+import { useParams, useRouter } from "next/navigation";
 import { useEffect } from "react";
-import { useMembersRegisterContext, MembersRegisterStep } from "../../../../../hooks/use-members-register-context";
+import { useMembersRegisterContext, MembersRegisterStep } from "../../../../../domains/patients/hooks/use-members-register-context";
 import PersonalForm from "../../../create/personal/page";
 import KinshipsForm from "../../../create/kinships/page";
 import AddressForm from "../../../create/address/page";
 import ResponsibleForm from "../../../create/responsible/page";
 import ProfileForm from "../../../create/profile/page";
-import AdditionalsForm from "../../../create/additional/page"; 
+import AdditionalsForm from "../../../create/additional/page";
 
 export default function EditPatientPage() {
   const { state: { step } } = useMembersRegisterContext();
-  const router = useRouter(); 
+  const router = useRouter();
   const { id, step: stepFromUrl } = useParams();
 
   useEffect(() => {
@@ -26,7 +26,7 @@ export default function EditPatientPage() {
         {step === MembersRegisterStep.PERSONAL && <PersonalForm />}
         {step === MembersRegisterStep.KINSHIPS && <KinshipsForm />}
         {step === MembersRegisterStep.ADDRESS && <AddressForm />}
-        {step === MembersRegisterStep.ADDITIONALS && <AdditionalsForm />} 
+        {step === MembersRegisterStep.ADDITIONALS && <AdditionalsForm />}
         {step === MembersRegisterStep.GUARDIAN && <ResponsibleForm />}
         {step === MembersRegisterStep.PROFILE && <ProfileForm />}
       </>
