+++ apps/apae/src/app/patients/create/responsible/page.tsx
@@ -12,15 +12,15 @@ import { Input } from "@/components/ui/input";
 import {
   MembersRegisterStep,
   useMembersRegisterContext,
-} from "@/hooks/use-members-register-context";
+} from "@/domains/patients/hooks/use-members-register-context";
 import { formatCEP, capitalizeFirst } from "@/lib/formats";
-import { Guardian, GuardianData } from "@/schemas/member-schemas";
+import { Guardian, GuardianData } from "@/domains/patients/schemas/member-schemas";
 import { EditGuardian } from "@/schemas/edit-member-schemas";
 import { zodResolver } from "@hookform/resolvers/zod";
-import React, { useState, useEffect } from "react"; 
+import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { handleBackendValidationErrors } from "@/lib/utils/form-errors";
-import { usePathname } from "next/navigation"; 
+import { usePathname } from "next/navigation";
 import { formatPhone } from "@/lib/formats";
 import { Checkbox } from "@/components/ui/checkbox";
 import z from "zod";
@@ -36,7 +36,7 @@ export default function MembersRegisterGuardianPage() {
 
   const pathname = usePathname();
   const isEditing = pathname.includes("/edit");
-  const currentSchema = isEditing ? EditGuardian : Guardian; 
+  const currentSchema = isEditing ? EditGuardian : Guardian;
 
   const form = useForm<z.infer<typeof Guardian>>({
     mode: "onBlur",
@@ -53,7 +53,7 @@ export default function MembersRegisterGuardianPage() {
   const isNoNumber = form.watch("address.noNumber");
 
   useEffect(() => {
-    if (isNoNumber){
+    if (isNoNumber) {
       form.setValue("address.number", "SN");
       form.clearErrors("address.number");
     } else {
@@ -64,7 +64,7 @@ export default function MembersRegisterGuardianPage() {
   }, [isNoNumber, form]);
 
   useEffect(() => {
-      if (guardian && guardian.name !== "") { 
+    if (guardian && guardian.name !== "") {
 
       form.reset({
         ...guardian,
@@ -91,10 +91,10 @@ export default function MembersRegisterGuardianPage() {
       setGuardianData(dataToSave);
       setStep(MembersRegisterStep.ADDRESS);
     } catch (error: unknown) {
-        const err = error as { response?: { data?: Record<string, string[]> } };
-        if (err.response?.data) {
-            handleBackendValidationErrors(err.response.data, form.setError);
-        }
+      const err = error as { response?: { data?: Record<string, string[]> } };
+      if (err.response?.data) {
+        handleBackendValidationErrors(err.response.data, form.setError);
+      }
     } finally {
       setIsLoading(false);
     }
@@ -233,7 +233,7 @@ export default function MembersRegisterGuardianPage() {
                         <label
                           htmlFor="noNumber"
                           className="text-[10px] font-bold uppercase text-slate-500 cursor-pointer select-none">
-                            Sem número?
+                          Sem número?
                         </label>
                       </div>
                     )}
