+++ apps/apae/src/types/form.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { EditAddress, EditPersonal, EditAdditionals, EditGuardian, EditProfile } from "@/schemas/edit-member-schemas";
-import { Address, Personal, Additionals, Guardian, Profile, Kinships } from "@/schemas/member-schemas";
+import { Address, Personal, Additionals, Guardian, Profile, Kinships } from "@/domains/patients/schemas/member-schemas";
 
 export type EditAddressFormValues = z.infer<typeof EditAddress>;
 export type EditPersonalFormValues = z.infer<typeof EditPersonal>;
